requirements.txt
streamlit
requests
pdfplumber
python-docx
fpdf
